Consolidate libs/backend CVR helpers
Under a flat src/cast_vote_records/ directory
