+++ libs/backend/src/ballot_package/index.ts
@@ -1,2 +1,3 @@
+/* istanbul ignore file */
 export * from './ballot_package_io';
 export * from './test_utils';
