+++ libs/backend/src/cast_vote_records/build_cast_vote_record.ts
