+++ libs/backend/src/cast_vote_records/build_report_metadata.ts
