+++ libs/backend/src/cast_vote_records/canonicalize.test.ts
@@ -10,7 +10,7 @@ import {
   interpretedBmdPage,
   interpretedHmpbPage1,
   interpretedHmpbPage2,
-} from '../../../test/fixtures/interpretations';
+} from '../../test/fixtures/interpretations';
 
 const blankPage: BlankPage = {
   type: 'BlankPage',
