+++ libs/backend/src/cast_vote_records/canonicalize.ts
