+++ libs/backend/src/cast_vote_records/export.ts
@@ -1,5 +1,4 @@
 /* istanbul ignore file */
-
 import path from 'path';
 import {
   computeSingleCastVoteRecordHash,
@@ -36,9 +35,9 @@ import {
   SCANNER_RESULTS_FOLDER,
 } from '@votingworks/utils';
 
-import { ExportDataError, Exporter } from '../../exporter';
-import { Usb as LegacyUsb } from '../../mock_usb';
-import { SCAN_ALLOWED_EXPORT_PATTERNS, VX_MACHINE_ID } from '../globals';
+import { ExportDataError, Exporter } from '../exporter';
+import { Usb as LegacyUsb } from '../mock_usb';
+import { SCAN_ALLOWED_EXPORT_PATTERNS, VX_MACHINE_ID } from '../scan/globals';
 import { buildCastVoteRecord as baseBuildCastVoteRecord } from './build_cast_vote_record';
 import { buildCastVoteRecordReportMetadata as baseBuildCastVoteRecordReportMetadata } from './build_report_metadata';
 import {
