+++ libs/backend/src/cast_vote_records/helpers.ts
