+++ libs/backend/src/cast_vote_records/option_map.test.ts
