+++ libs/backend/src/index.ts
@@ -1,10 +1,9 @@
 /* istanbul ignore file */
+export * from './ballot_package';
+export * from './cast_vote_records';
 export * from './exporter';
 export * from './get_usb_drives';
+export * from './list_directory';
+export * from './mock_usb';
 export * from './scan';
 export * from './split';
-export * from './list_directory';
-export * from './cast_vote_record_helpers';
-export * from './cast_vote_record_report_import';
-export * from './ballot_package';
-export * from './mock_usb';
