+++ libs/backend/src/scan/index.ts
@@ -1,7 +1,2 @@
-export * from './cast_vote_records/page_layouts';
-export * from './cast_vote_records/build_cast_vote_record';
-export * from './cast_vote_records/build_report_metadata';
-export * from './cast_vote_records/legacy_export';
-export * from './cast_vote_records/canonicalize';
-export * from './cast_vote_records/export';
+/* istanbul ignore file */
 export { VX_MACHINE_ID } from './globals';
